+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 Change Log
 ==========
 
+Version 2.0.7 *(10 May, 2018)*
+-------------------------------------------
+Update for Android support library version
+Update for CleverTap Android SDK v 3.1.9
+Update for CleverTap iOS SDK v 3.1.7
+Support for Android O Notification Channels with custom sound
+New APIs for GDPR compliance
+Adds Android support for recordScreenView API
+
 Version 2.0.5 *(04 January, 2018)*
 -------------------------------------------
 Update for CleverTap Android SDK v 3.1.8